requirements.txt
line-bot-sdk 
flask 
SnowNLP 
emoji 
pyshorteners 
python-dotenv